move general logit function to calo_utils
