+++ experiments/calo_utils/transform_utils.py
@@ -1,0 +1,11 @@
+import torch
+
+
+def logit(array, alpha=1.0e-6, inv=False):
+    if inv:
+        z = torch.sigmoid(array)
+        z = (z - alpha) / (1 - 2 * alpha)
+    else:
+        z = array * (1 - 2 * alpha) + alpha
+        z = torch.logit(z)
+    return z
