+++ experiments/calochallenge/transforms.py
@@ -5,17 +5,8 @@ import numpy as np
 import torch
 import torch.nn.functional as F
 
+from experiments.calo_utils.transform_utils import logit
 from experiments.calo_utils.ugr_evaluation import XMLHandler
-
-
-def logit(array, alpha=1.0e-6, inv=False):
-    if inv:
-        z = torch.sigmoid(array)
-        z = (z - alpha) / (1 - 2 * alpha)
-    else:
-        z = array * (1 - 2 * alpha) + alpha
-        z = torch.logit(z)
-    return z
 
 
 class GlobalStandardizeFromFile:
